// src/app/modules/admin/predial-inspection/results-management/interfaces/caracteristica.interface.ts
export interface ICaracteristica{
    categoriaElectrica: string;
    piso: string;
    estadoConserva: string;
    anioConstruccion: number;
    catergoriaTecho: string;
    longitudFrente: number;
    categoriaMuroColumna: string;
    catergoriaPuertaVentana: string;
    arancel: number;
    materialPred: string;
    codTit: string;
    categoriaRevestimiento: string;
    areaTerreno: number;
    clasificacionPred: string;
    catergoriaPiso: string;
    catergoriaBano: string;
    areaConstruida: number;
}